alpha 0.1.1 timers refactoring

+++ include/pins.h
@@ -38,20 +38,20 @@ class Beeper : public Pin
 private:
   bool isActivated;
   uint8_t repeat;
-  uint32_t start;
-  uint32_t end;
-  uint32_t pause;
+  uint16_t start;
+  uint16_t end;
+  uint16_t pause;
 
 public:
   using Pin::Pin;
   void setOn();
   void setOff();
-  void setBeep(uint32_t startTime, uint32_t duration, uint8_t repeatCount = 1, uint32_t pauseDuration = 0); // defaults: repeatCount = 1, pauseDuration = duration
+  void setBeep(uint16_t startTime, uint16_t duration, uint8_t repeatCount = 1, uint16_t pauseDuration = 0); // defaults: repeatCount = 1, pauseDuration = duration
   void beep();
   void beepOnce();
   void beepTwice();
-  uint32_t getStart();
-  uint32_t getEnd();
+  uint16_t getStart();
+  uint16_t getEnd();
   bool isOn();
 };
